Moved dateformating to bean, updated generic loadbar for document
